requirements.txt
arabic-reshaper==2.1.3
asgiref==3.6.0
asn1crypto==1.5.1
blinker==1.4
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
click==8.0.4
colorama==0.4.4
crispy-bootstrap4==2022.1
crispy-bootstrap5==0.6
cryptography==38.0.1
cssselect2==0.6.0
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.4
dj-database-url==2.0.0
Django==4.2
django-active-link==0.1.8
django-appconf==1.0.5
django-autocomplete-light==3.9.4
django-ckeditor==6.5.1
django-countries==7.5.1
django-crispy-forms==2.0
django-discover-runner==1.0
django-environ==0.10.0
django-extensions==3.2.1
django-file-form==3.4.1
django-filter==22.1
django-heroku==0.3.1
django-import-export==2.8.0
django-js-asset==2.0.0
django-model-utils==4.3.1
django-phonenumber-field==7.0.2
django-role-permissions==3.1.1
django-select2==7.10.1
django-taggit==3.0.0
Django-Verify-Email==2.0.3
django-widget-tweaks==1.4.12
et-xmlfile==1.1.0
filelock==3.6.0
Flask==2.0.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.1.2
gunicorn==21.0.0
html5lib==1.1
idna==3.4
itsdangerous==2.1.0
Jinja2==3.0.3
lxml==4.9.1
MarkupPy==1.14
MarkupSafe==2.1.0
mysqlclient==2.1.1
odfpy==1.4.1
openpyxl==3.0.10
oscrypto==1.3.0
packaging==23.1
password-strength==0.0.3.post2
phonenumberslite==8.13.10
Pillow==9.5.0
platformdirs==2.5.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
pydot==1.4.2
pyHanko==0.13.2
pyhanko-certvalidator==0.19.5
pyparsing==3.0.9
PyPDF3==1.0.6
python-bidi==0.4.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.11
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.31
sqlparse==0.4.3
style==1.1.0
svglib==1.4.1
tablib==3.2.1
tinycss2==1.1.1
tqdm==4.64.1
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.2
update==0.0.1
uritools==4.0.0
urllib3==1.26.12
virtualenv==20.14.0
webencodings==0.5.1
Werkzeug==2.0.3
whitenoise==6.5.0
xhtml2pdf==0.2.8
xlrd==2.0.1
xlwt==1.3.0

